requirements.lock.txt
Scrapy==2.13.3
itemloaders==1.3.2
lxml==6.0.1
cssselect==1.3.0
requests==2.32.4
tldextract==5.3.0

spacy==3.8.7
pysentimiento==0.7.3
nltk==3.9.1
numpy==2.3.2

transformers==4.55.4
accelerate==1.10.1

psycopg[binary]==3.2.9
SQLAlchemy==2.0.43
python-dotenv==1.0.1
pandas==2.3.2
matplotlib==3.10.5
typer==0.16.1

protobuf==6.32.0
regex==2025.7.34
pydantic==2.11.7
